Remove call to load legacy provider from libsm
Library code should *never* arbitrarily load the legacy provider in the
default application context. It may literally break or make an
application vulnerable by sneaking in unexpected algorithms.

This is not needed anymore given we will use an opensc specific context
that will not affect the rest of the application (or other libraries).

Signed-off-by: Simo Sorce <simo@redhat.com>